refactor code

+++ Entities/Transcript.cs
@@ -10,13 +10,13 @@ namespace Entities
         public int Id { get; set; }
         public int StudentId { get; set; }
         public int SubjectId { get; set; }
-        public float MidtermGradeI { get; set; }                                             
-        public float FinalGradeI { get; set; }
-        public float AverageGradeI { get; set; }
-        public float MidtermGradeII { get; set; }
-        public float FinalGradeII { get; set; }
-        public float AverageGradeII { get; set; }
-        public float FinalAverageGrade { get; set; }
+        public double MidtermGradeI { get; set; }
+        public double FinalGradeI { get; set; }
+        public double AverageGradeI { get; set; }
+        public double MidtermGradeII { get; set; }
+        public double FinalGradeII { get; set; }
+        public double AverageGradeII { get; set; }
+        public double FinalAverageGrade { get; set; }
         public string GradeType { get; set; }
 
         [ForeignKey("StudentId")]
